Refs #25 - Abstract setup of mock security context into controller test case.
Make controller test case abstract, since that is its intent.
